fix error in baseline

+++ trainer.py
@@ -20,7 +20,7 @@ class CustomTrainer(Trainer):
         outputs = model(**inputs)
 
         if labels is not None:
-            loss = custom_loss(outputs[0], labels)
+            loss = custom_loss(outputs[0], labels.long())
         else:
             # We don't use .loss here since the model may return tuples instead of ModelOutput.
             loss = outputs["loss"] if isinstance(outputs, dict) else outputs[0]
